requirements.txt
Django==1.6.1
SPARQLWrapper==1.5.2
South==0.8.4
amqp==1.4.4
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.16
blist==1.3.6
cassandra-driver==1.0.2
celery==3.1.10
distribute==0.6.24
django-debug-toolbar==1.0.1
django-environ==0.2.1
django-extensions==1.3.3
django-statsd-mozilla==0.3.10
django-vanilla-views==1.0.2
futures==2.1.6
honcho==0.5.0
html5lib==1.0b3
isodate==0.5.0
kombu==3.0.14
lz4==0.6.1
psycopg2==2.5.2
pyparsing==2.0.1
python-statsd==1.6.3
pytz==2014.1.1
rdflib==4.1.1
rdflib-django==0.3
scales==1.0.3
six==1.6.1
sqlparse==0.1.11
statsd==2.1.2
time-uuid==0.1.1
wsgiref==0.1.2
